Finish refactor

+++ api/scraper/login.py
@@ -1,5 +1,6 @@
 import logging
 import pathlib
+import pickle
 import secrets
 import shutil
 
@@ -7,7 +8,7 @@ from fastapi import HTTPException
 from selenium import webdriver
 from selenium.common.exceptions import TimeoutException, ElementNotInteractableException
 from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions as EC
+from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.ui import WebDriverWait
 
 from . import common
@@ -15,10 +16,52 @@ from .common import delete_session
 
 URL = "https://quest.pecs.uwaterloo.ca/psc/AS/ACADEMIC/SA/c/NUI_FRAMEWORK.PT_LANDINGPAGE.GBL"
 
+DUMMY_URL = "https://quest.pecs.uwaterloo.ca/nonexistent"
+
+PROFILE_PATH = f"{pathlib.Path(__file__).parent.parent}/profiles"
+
 
 class UserAuthenticationException(Exception):
     def __init__(self, message):
         super().__init__(message)
+
+
+def create_token() -> str:
+    """
+    Creates a new token for user
+    :return: token
+    """
+    return secrets.token_urlsafe(16)
+
+
+def dump_cookies(token: str, cookies: dict) -> None:
+    """
+    Utility function to dump cookies to file in profile path
+    :param token: User token
+    :param cookies: Dictionary of cookies
+    :return: None
+    """
+    with open(f"{PROFILE_PATH}/{token}/cookieJar.pkl", "wb+") as f:
+        pickle.dump(cookies, f)
+
+
+def load_cookies(token: str) -> None:
+    """
+    Utility function to load cookies from file
+    :param token: User token
+    :return: None
+    """
+    driver = common.driver_list[token]
+    try:
+        with open(f"{PROFILE_PATH}/{token}/cookieJar.pkl", "rb") as f:
+            cookies = pickle.load(f)
+    except FileNotFoundError:
+        logging.warning("No cookies found for %s", token)
+        return
+    driver.get(DUMMY_URL)
+    for cookie in cookies:
+        logging.debug("Adding cookie %s", cookie)
+        common.driver_list[token].add_cookie(cookie)
 
 
 def ini_driver(token: str, remember_me: bool) -> webdriver.Edge:
@@ -32,7 +75,7 @@ def ini_driver(token: str, remember_me: bool) -> webdriver.Edge:
     if not remember_me:
         options.add_argument("inprivate")
     else:
-        options.add_argument(f"user-data-dir={pathlib.Path(__file__).parent}/../profiles/{token}")
+        options.add_argument(f"user-data-dir={PROFILE_PATH}/{token}")
     options.add_experimental_option("detach", True)
     # options.add_argument("--headless=new")
     options.add_argument("--no-sandbox")
@@ -60,40 +103,38 @@ def recreate_session(token: str) -> None:
         logging.info("Recreating session for %s", token)
         driver = ini_driver(token, True)
 
+    load_cookies(token)
     driver.get(URL)
 
     if not common.verify_signed_on(token):
         raise UserAuthenticationException("Session expired, sign in again")
 
 
-def sign_in(user: str, credentials: str, remember_me: bool, token: str) -> str:
+def sign_in(user: str, credentials: str, remember_me: bool) -> [str, str]:
     """
     Sign in user and creates new driver instance if new user
-    :param token: newly generated token
-    :param user: uwaterloo email
+    :param user: uWaterloo email
     :param credentials: password
     :param remember_me: if session should be persistent
-    :return: duo auth code if required, '' if not
+    :return: [token, duo_auth_code] duo_auth_code may be None if not required
     """
-
     if user in common.known_users:
         logging.info("User %s already assigned token %s", user, common.known_users[user])
         token = common.known_users[user]
-    elif remember_me:
+        recreate_session(token)
+        return [token, '']
+
+    token = create_token()
+    if remember_me:
         common.known_users[user] = token
 
-    if token in common.driver_list:
-        driver = common.driver_list[token]
-    else:
-        driver = ini_driver(token, remember_me)
+    driver = ini_driver(token, remember_me)
+
     # navigate to sign in form
     driver.get(URL)
 
-    if common.verify_signed_on(token):  # user already signed in
-        return ''
-
     try:
-        WebDriverWait(driver, timeout=10).until(EC.title_is("Sign In"))
+        WebDriverWait(driver, timeout=10).until(ec.title_is("Sign In"))
         try:
             logging.info("Signing in as %s", user)
             username = WebDriverWait(driver, timeout=10).until(lambda d: d.find_element(By.ID, 'userNameInput'))
@@ -111,30 +152,36 @@ def sign_in(user: str, credentials: str, remember_me: bool, token: str) -> str:
 
     if driver.title == "Homepage":  # already signed in
         logging.info("DUO Auth passed by cookie")
-        return ''
+        return [token, '']
     else:
-        WebDriverWait(driver, timeout=10).until(EC.presence_of_element_located((By.CLASS_NAME, "verification-code")))
+        WebDriverWait(driver, timeout=10).until(ec.presence_of_element_located((By.CLASS_NAME, "verification-code")))
         duo_auth_code = driver.find_element(By.CLASS_NAME, 'verification-code').text
         logging.info(f"Parsed DUO Auth code: {duo_auth_code}")
-        return duo_auth_code
+        return [token, duo_auth_code]
 
 
-def duo_auth(token: str, remember_me) -> str:
+def duo_auth(token: str, remember_me: bool) -> None:
+    """
+    Completes duo auth login step
+    :param token: User token
+    :param remember_me: If user should be remembered
+    :return: None
+    """
     driver = common.driver_list[token]
     logging.info("DUO Auth required. Waiting for user interaction...")
     try:
         wait = WebDriverWait(driver, timeout=120)
         if remember_me:
-            wait.until(EC.element_to_be_clickable((By.ID, "trust-browser-button"))).click()
+            wait.until(ec.element_to_be_clickable((By.ID, "trust-browser-button"))).click()
         else:
-            wait.until(EC.element_to_be_clickable((By.ID, "dont-trust-browser-button"))).click()
+            wait.until(ec.element_to_be_clickable((By.ID, "dont-trust-browser-button"))).click()
         # wait until duo auth is passed
-        WebDriverWait(driver, timeout=60).until(EC.title_is("Homepage"))
+        WebDriverWait(driver, timeout=60).until(ec.title_is("Homepage"))
         logging.info("Sign in successful for %s", token)
-        return token
+
+        dump_cookies(token, driver.get_cookies())
     except TimeoutException:
         logging.error("Duo Auth timed out")
-        sign_out(token)
         raise UserAuthenticationException("Duo Auth timed out")
 
 
@@ -145,7 +192,11 @@ def sign_out(token: str) -> str:
     :return: token that was signed out
     """
     logging.info("Signing out user %s", token)
-    delete_session(token, True)
+    delete_session(token)
     # delete profile folder
-    shutil.rmtree(f"profiles/{token}")
+    user_to_delete = {k: v for k, v in common.known_users.items() if v == token}.keys()
+    del common.known_users[user_to_delete]
+    if user_to_delete is not None:
+        shutil.rmtree(f"profiles/{token}")
+
     return token
